Reorganize railway rules for better parallel track rendering.
In the US, the horizontal distance between parallel main tracks is around
15 to 18 feet (depending on the speed of the trains and the age of the
track).  This works out as 2 to 2.5 pixels at zoom 16.  The existing
rendering had 4 pixel railway casings and 2.5 pixel fills, which meant
that they almost always overlapped for parallel tracks, which looked a
little bad.

I've gone back to the older TopOSM rendering approach of drawing all the
outlines first and then doing the fills, which eliminates the problem of
adjacent tracks' outlines infringing on their neighbor's fills.  I also
narrowed the fill a little so that there will be a little separation
between fills at zoom 16.

Railyards are a bit of a mess now (well, they were before, too, I think,
but now they're worse) with white fill everywhere, but that's another
changeset.